Refactor CompanionPage into smaller components and custom hook
- Extracted state and logic to `useChat` hook.
- Created `ChatHeader`, `ChatMessageList`, `ChatInput` components in `src/app/dashboard/chat/_components/`.
- Moved types to `src/app/dashboard/chat/types.ts`.
- Moved constants to `src/app/dashboard/chat/constants.ts`.
- Moved utils to `src/app/dashboard/chat/utils.ts`.
- Removed unused `formatHistoryTime` function.
- Verified functionality with manual test page and playwright.
- Verified build and existing tests.
